@@ -18,95 +18,82 @@
   "main": "main.js",
   "private": true,
   "scripts": {
-    "ap_download:windows": "pwsh -NoProfile -ExecutionPolicy ByPass -File ./dev/script_ap.ps1",
-    "ap_download:linux": "npx pwsh -NoProfile -ExecutionPolicy ByPass -File ./dev/script_ap.ps1",
-    "copy_ap:windows": "pwsh -NoProfile -ExecutionPolicy ByPass -File ./dev/azure_setup.ps1",
-    "copy_ap:linux": "npx pwsh -NoProfile -ExecutionPolicy ByPass -File ./dev/azure_setup.ps1",
     "postinstall": "electron-builder install-app-deps",
     "ng": "ng",
-    "start": "npm-run-all -p ng:serve electron:serve",
-    "build": "npm run electron:serve-tsc && ng build",
+    "start": "npm-run-all -p electron:serve ng:serve",
+    "build": "npm run electron:serve-tsc && ng build --base-href ./",
     "build:dev": "npm run build -- -c dev",
     "build:prod": "npm run build -- -c production",
-    "dependencies": "chmod +x ./dev/linux_dependencies.sh && ./dev/linux_dependencies.sh",
-    "ng:serve": "ng serve",
-    "ng:serve:web": "ng serve -c web -o",
-    "electron:serve-tsc": "tsc -p tsconfig-serve.json",
-    "electron:serve": "wait-on http-get://localhost:4200/ && npm run electron:serve-tsc && electron . --serve",
-    "electron:local": "npm run build:prod && electron .",
-    "electron:linux": "npm run build:prod && electron-builder build --linux --publish=never",
-    "electron:windows": "npm run build:prod && electron-builder build --windows --publish=never",
-    "electron:mac": "npm run build:prod && electron-builder build --mac --publish=never",
-    "test": "ng test",
-    "e2e": "npm run build:prod && mocha --timeout 300000 --require ts-node/register e2e/**/*.spec.ts",
+    "ng:serve": "ng serve -c web -o",
+    "electron:serve-tsc": "tsc -p tsconfig.serve.json",
+    "electron:serve": "wait-on tcp:4200 && npm run electron:serve-tsc && npx electron . --serve",
+    "electron:local": "npm run build:prod && npx electron .",
+    "electron:build": "npm run build:prod && electron-builder build --publish=never",
+    "test": "ng test --watch=false",
+    "test:watch": "ng test",
+    "e2e": "npm run build:prod && cross-env TS_NODE_PROJECT='e2e/tsconfig.e2e.json' mocha --timeout 300000 --require ts-node/register e2e/**/*.e2e.ts",
     "version": "conventional-changelog -i CHANGELOG.md -s -r 0 && git add CHANGELOG.md",
     "lint": "ng lint"
   },
+  "dependencies": {},
   "devDependencies": {
-    "@angular-builders/custom-webpack": "8.2.0",
-    "@angular-devkit/build-angular": "0.803.23",
-    "@angular-eslint/builder": "0.0.1-alpha.17",
-    "@angular/cli": "8.3.21",
-    "@angular/common": "8.2.14",
-    "@angular/compiler": "8.2.14",
-    "@angular/compiler-cli": "8.2.14",
-    "@angular/core": "8.2.14",
-    "@angular/forms": "8.2.14",
-    "@angular/language-service": "8.2.14",
-    "@angular/platform-browser": "8.2.14",
-    "@angular/platform-browser-dynamic": "8.2.14",
-    "@angular/router": "8.2.14",
-    "@ngx-translate/core": "11.0.1",
-    "@ngx-translate/http-loader": "4.0.0",
-    "@types/jasmine": "3.3.16",
+    "@angular-builders/custom-webpack": "11.0.0",
+    "@angular-devkit/build-angular": "0.1102.0",
+    "@angular-eslint/builder": "1.2.0",
+    "@angular-eslint/eslint-plugin": "1.2.0",
+    "@angular-eslint/eslint-plugin-template": "1.2.0",
+    "@angular-eslint/schematics": "1.2.0",
+    "@angular-eslint/template-parser": "1.2.0",
+    "@angular/cli": "11.2.0",
+    "@angular/common": "11.2.0",
+    "@angular/compiler": "11.2.0",
+    "@angular/compiler-cli": "11.2.0",
+    "@angular/core": "11.2.0",
+    "@angular/forms": "11.2.0",
+    "@angular/language-service": "11.2.4",
+    "@angular/platform-browser": "11.2.0",
+    "@angular/platform-browser-dynamic": "11.2.0",
+    "@angular/router": "11.2.0",
+    "@types/jasmine": "3.6.3",
     "@types/jasminewd2": "2.0.8",
-    "@types/js-yaml": "3.12.1",
-    "@types/mocha": "5.2.7",
-    "@types/node": "12.6.9",
-    "@typescript-eslint/eslint-plugin": "2.8.0",
-    "@typescript-eslint/parser": "2.8.0",
+    "@types/mocha": "8.2.0",
+    "@types/node": "12.12.6",
+    "@typescript-eslint/eslint-plugin": "4.15.0",
+    "@typescript-eslint/eslint-plugin-tslint": "4.15.0",
+    "@typescript-eslint/parser": "4.15.0",
     "chai": "4.2.0",
-    "codelyzer": "5.1.2",
-    "conventional-changelog-cli": "2.0.32",
-    "core-js": "3.1.4",
-    "electron": "7.1.7",
-    "electron-builder": "21.2.0",
+    "conventional-changelog-cli": "2.1.1",
+    "core-js": "3.6.5",
+    "cross-env": "7.0.3",
+    "electron": "11.2.0",
+    "electron-builder": "22.9.1",
     "electron-reload": "1.5.0",
-    "eslint": "6.6.0",
-    "eslint-plugin-import": "2.18.2",
-    "jasmine-core": "3.4.0",
-    "jasmine-spec-reporter": "4.2.1",
-    "karma": "4.2.0",
-    "karma-coverage-istanbul-reporter": "2.1.1",
-    "karma-electron": "6.3.0",
-    "karma-jasmine": "2.0.1",
-    "karma-jasmine-html-reporter": "1.4.2",
-    "mocha": "6.2.2",
+    "eslint": "7.20.0",
+    "eslint-plugin-import": "2.22.1",
+    "eslint-plugin-jsdoc": "31.6.1",
+    "eslint-plugin-prefer-arrow": "1.2.3",
+    "jasmine-core": "3.6.0",
+    "jasmine-spec-reporter": "6.0.0",
+    "karma": "6.1.1",
+    "karma-coverage-istanbul-reporter": "3.0.3",
+    "karma-electron": "6.3.3",
+    "karma-jasmine": "4.0.1",
+    "karma-jasmine-html-reporter": "1.5.4",
+    "mocha": "8.2.1",
     "npm-run-all": "4.1.5",
-    "rxjs": "6.5.3",
-    "rxjs-tslint-rules": "4.28.0",
-    "spectron": "9.0.0",
-    "ts-node": "8.3.0",
-    "tslint": "5.20.1",
-    "tslint-angular": "3.0.2",
-    "typescript": "3.5.3",
-    "wait-on": "3.3.0",
-    "webdriver-manager": "12.1.7",
-    "zone.js": "0.9.1"
+    "rxjs": "6.6.3",
+    "spectron": "13.0.0",
+    "ts-node": "9.1.1",
+    "tslib": "2.1.0",
+    "typescript": "4.0.5",
+    "wait-on": "5.2.1",
+    "webdriver-manager": "12.1.8",
+    "zone.js": "0.10.3"
   },
   "engines": {
-    "node": ">=10.9.0"
+    "node": ">=10.13.0"
   },
-  "dependencies": {
-    "@fortawesome/angular-fontawesome": "0.5.0",
-    "@fortawesome/fontawesome-svg-core": "1.2.26",
-    "@fortawesome/free-brands-svg-icons": "5.12.0",
-    "@fortawesome/free-regular-svg-icons": "5.12.0",
-    "@fortawesome/free-solid-svg-icons": "5.12.0",
-    "@ng-bootstrap/ng-bootstrap": "5.1.5",
-    "@swimlane/ngx-datatable": "16.0.3",
-    "bootstrap": "4.4.1",
-    "immutable": "4.0.0-rc.12",
-    "js-yaml": "3.13.1"
-  }
+  "browserslist": [
+    "chrome 83"
+  ]
 }
