@@ -3,26 +3,22 @@
  */
 
 module.exports = (config, options) => {
-  config.target = "electron-renderer";
-  if (options.customWebpackConfig.target) {
-    config.target = options.customWebpackConfig.target;
-  } else if (options.fileReplacements) {
-    for (let fileReplacement of options.fileReplacements) {
-      if (fileReplacement.replace !== "src/environments/environment.ts") {
-        continue;
-      }
+    config.target = 'electron-renderer';
 
-      let fileReplacementParts = fileReplacement["with"].split(".");
-      if (
-        ["dev", "prod", "test", "electron-renderer"].indexOf(
-          fileReplacementParts[1]
-        ) < 0
-      ) {
-        config.target = fileReplacementParts[1];
-      }
-      break;
+
+    if (options.fileReplacements) {
+        for(let fileReplacement of options.fileReplacements) {
+            if (fileReplacement.replace !== 'src/environments/environment.ts') {
+                continue;
+            }
+
+            let fileReplacementParts = fileReplacement['with'].split('.');
+            if (fileReplacementParts.length > 1 && ['web'].indexOf(fileReplacementParts[1]) >= 0) {
+                config.target = 'web';
+            }
+            break;
+        }
     }
-  }
 
-  return config;
-};
+    return config;
+}
