@@ -1,9 +1,9 @@
-const Application = require("spectron").Application;
-const electronPath = require("electron"); // Require Electron from the binaries included in node_modules.
-const path = require("path");
+const Application = require('spectron').Application;
+const electronPath = require('electron'); // Require Electron from the binaries included in node_modules.
+const path = require('path');
 
 export default function setup(): void {
-  beforeEach(async function() {
+  beforeEach(async function () {
     this.app = new Application({
       // Your electron path can be any binary
       // i.e for OSX an example path could be '/Applications/MyApp.app/Contents/MacOS/MyApp'
@@ -23,19 +23,16 @@ export default function setup(): void {
 
       // The following line tells spectron to look and use the main.js file
       // and the package.json located 1 level above.
-      args: [path.join(__dirname, "..")],
+      args: [path.join(__dirname, '..')],
       webdriverOptions: {}
     });
-    await this.app.start();
-    const browser = this.app.client;
-    await browser.waitUntilWindowLoaded();
 
-    browser.timeouts("script", 15000);
+    await this.app.start();
   });
 
-  afterEach(function() {
+  afterEach(async function () {
     if (this.app && this.app.isRunning()) {
-      return this.app.stop();
+      await this.app.stop();
     }
   });
 }
