@@ -4,7 +4,7 @@ import * as url from "url";
 
 let win: BrowserWindow = null;
 const args = process.argv.slice(1),
-  serve = args.some(val => val === "--serve");
+  serve = args.some((val) => val === "--serve");
 
 function createWindow(): BrowserWindow {
   const electronScreen = screen;
@@ -18,13 +18,17 @@ function createWindow(): BrowserWindow {
     height: size.height,
     webPreferences: {
       nodeIntegration: true,
-      allowRunningInsecureContent: serve ? true : false
-    }
+      allowRunningInsecureContent: serve ? true : false,
+      contextIsolation: false, // false if you want to run 2e2 test with Spectron
+      enableRemoteModule: true, // true if you want to run 2e2 test  with Spectron or use remote module in renderer context (ie. Angular)
+    },
   });
 
   if (serve) {
+    win.webContents.openDevTools();
+
     require("electron-reload")(__dirname, {
-      electron: require(`${__dirname}/node_modules/electron`)
+      electron: require(`${__dirname}/node_modules/electron`),
     });
     win.loadURL("http://localhost:4200");
   } else {
@@ -32,13 +36,9 @@ function createWindow(): BrowserWindow {
       url.format({
         pathname: path.join(__dirname, "dist/index.html"),
         protocol: "file:",
-        slashes: true
+        slashes: true,
       })
     );
-  }
-
-  if (serve) {
-    win.webContents.openDevTools();
   }
 
   // Emitted when the window is closed.
@@ -56,7 +56,8 @@ try {
   // This method will be called when Electron has finished
   // initialization and is ready to create browser windows.
   // Some APIs can only be used after this event occurs.
-  app.on("ready", createWindow);
+  // Added 400 ms to fix the black background issue while using transparent window. More detais at https://github.com/electron/electron/issues/15947
+  app.on("ready", () => setTimeout(createWindow, 400));
 
   // Quit when all windows are closed.
   app.on("window-all-closed", () => {
