@@ -1,32 +1,23 @@
-import { Component, OnInit } from "@angular/core";
-import { TranslateService } from "@ngx-translate/core";
+import { Component } from "@angular/core";
+import { AppConfig } from "../environments/environment";
+import { ElectronService } from "./core/services";
 
 @Component({
   selector: "app-root",
   templateUrl: "./app.component.html",
-  styleUrls: ["./app.component.scss"]
+  styleUrls: ["./app.component.scss"],
 })
-export class AppComponent implements OnInit {
-  constructor(private translate: TranslateService) {
-    this.translate.setDefaultLang("en");
-    // console.log("AppConfig", AppConfig);
+export class AppComponent {
+  constructor(private electronService: ElectronService) {
+    console.log("AppConfig", AppConfig);
 
-    // if (electronService.isElectron) {
-    //   console.log(process.env);
-    //   console.log("Mode electron");
-    //   console.log("Electron ipcRenderer", electronService.ipcRenderer);
-    //   console.log("NodeJS childProcess", electronService.childProcess);
-    // } else {
-    //   console.log("Mode web");
-    // }
-  }
-
-  ngOnInit(): void {
-    // TODO Remove this once a solution is found to the ChildProcess bug
-    const started = sessionStorage.getItem("started");
-    if (!started) {
-      sessionStorage.setItem("started", "true");
-      window.location.reload();
+    if (electronService.isElectron) {
+      console.log(process.env);
+      console.log("Run in electron");
+      console.log("Electron ipcRenderer", this.electronService.ipcRenderer);
+      console.log("NodeJS childProcess", this.electronService.childProcess);
+    } else {
+      console.log("Run in browser");
     }
   }
 }
