@@ -1,7 +1,7 @@
 // Karma configuration file, see link for more information
 // https://karma-runner.github.io/0.13/config/configuration-file.html
 
-module.exports = function(config) {
+module.exports = function (config) {
   config.set({
     basePath: "",
     frameworks: ["jasmine", "@angular-devkit/build-angular"],
@@ -10,36 +10,34 @@ module.exports = function(config) {
       require("karma-electron"),
       require("karma-jasmine-html-reporter"),
       require("karma-coverage-istanbul-reporter"),
-      require("@angular-devkit/build-angular/plugins/karma")
+      require("@angular-devkit/build-angular/plugins/karma"),
     ],
     client: {
-      clearContext: false // leave Jasmine Spec Runner output visible in browser
+      clearContext: false, // leave Jasmine Spec Runner output visible in browser
     },
     coverageIstanbulReporter: {
       dir: require("path").join(__dirname, "../coverage"),
       reports: ["html", "lcovonly"],
-      fixWebpackSourcePaths: true
+      fixWebpackSourcePaths: true,
     },
     reporters: ["progress", "kjhtml"],
     port: 9876,
     colors: true,
     logLevel: config.LOG_INFO,
-    autoWatch: true,
     browsers: ["AngularElectron"],
-    singleRun: true,
     customLaunchers: {
       AngularElectron: {
         base: "Electron",
+        flags: ["--remote-debugging-port=9222"],
         browserWindowOptions: {
           webPreferences: {
             nodeIntegration: true,
-            allowRunningInsecureContent: true
-          }
-        }
-      }
+            nodeIntegrationInSubFrames: true,
+            allowRunningInsecureContent: true,
+            enableRemoteModule: true,
+          },
+        },
+      },
     },
-    client: {
-      useIframe: false
-    }
   });
 };
