@@ -9,6 +9,6 @@ if (AppConfig.production) {
 
 platformBrowserDynamic()
   .bootstrapModule(AppModule, {
-    preserveWhitespaces: false
+    preserveWhitespaces: false,
   })
-  .catch(err => console.error(err));
+  .catch((err) => console.error(err));
