@@ -1,2 +1,2 @@
-import "core-js/es/reflect";
-import "zone.js/dist/zone";
+import 'core-js/es/reflect';
+import 'zone.js/dist/zone';
